api/requirements.txt
django
djangorestframework
python-dotenv
requests
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
pytz